Refactor subscribe now and status pages into separate widgets
