+++ gui/packages/ubuntupro/lib/app.dart
@@ -11,6 +11,7 @@ import 'core/agent_api_client.dart';
 import 'launch_agent.dart';
 import 'pages/startup/agent_monitor.dart';
 import 'pages/startup/startup_page.dart';
+import 'pages/subscribe_now/subscribe_now_page.dart';
 import 'pages/subscription_status/subscription_status_page.dart';
 import 'routes.dart';
 
@@ -35,6 +36,8 @@ class Pro4WSLApp extends StatelessWidget {
               return Wizard(
                 routes: {
                   Routes.startup: const WizardRoute(builder: buildStartup),
+                  Routes.subscribeNow:
+                      const WizardRoute(builder: SubscribeNowPage.create),
                   Routes.subscriptionStatus: WizardRoute(
                     builder: SubscriptionStatusPage.create,
                     onLoad: (_) async {
