+++ gui/packages/ubuntupro/lib/routes.dart
@@ -1,5 +1,6 @@
 abstract class Routes {
   static const startup = '/startup';
+  static const subscribeNow = '/subscribe-now';
   static const configureLandscape = '/configure-landscape';
   static const subscriptionStatus = '/subscription-status';
 }
