+++ gui/packages/ubuntupro/test/pages/subcribe_now/subscribe_now_page_test.mocks.dart
@@ -1,5 +1,5 @@
 // Mocks generated by Mockito 5.4.4 from annotations
-// in ubuntupro/test/pages/subscription_status/subscribe_now_page_test.dart.
+// in ubuntupro/test/pages/subcribe_now/subscribe_now_page_test.dart.
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
@@ -11,8 +11,7 @@ import 'package:mockito/mockito.dart' as _i1;
 import 'package:p4w_ms_store/p4w_ms_store.dart' as _i8;
 import 'package:ubuntupro/core/agent_api_client.dart' as _i2;
 import 'package:ubuntupro/core/pro_token.dart' as _i7;
-import 'package:ubuntupro/pages/subscription_status/subscription_status_model.dart'
-    as _i5;
+import 'package:ubuntupro/pages/subscribe_now/subscribe_now_model.dart' as _i5;
 
 // ignore_for_file: type=lint
 // ignore_for_file: avoid_redundant_argument_values
